% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AS.R
\name{AS}
\alias{AS}
\title{Adaptive selection method}
\usage{
AS(R, lambda)
}
\arguments{
\item{R}{A numeric vector of R-values}

\item{lambda}{Numeric, the tuning parameter}
}
\value{
The number of effective populations K
}
\description{
This function performs the AS method to estimate the number of effective populations K
}
