% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAS.R
\name{RAS}
\alias{RAS}
\title{Randomised Adaptive selection method}
\usage{
RAS(R, lambda)
}
\arguments{
\item{R}{A numeric vector of R-values}

\item{lambda}{Numeric, the tuning parameter}
}
\value{
The number of effective populations K
}
\description{
This function performs the AS method to estimate the number of effective populations K
}
